implement masknet core model as a module component
