+++ tzrec/protos/models/rank_model.proto
@@ -37,12 +37,5 @@ message DLRM {
 
 
 message MaskNet {
-    // number of mask blocks
-    required uint32 n_mask_blocks = 1;
-    // mask block
-    required MaskBlock mask_block = 2;
-    required MLP top_mlp = 3;
-    // use parallel or serial mask blocks
-    optional bool use_parallel = 4 [default=true];
-
+    required MaskNetModule mask_net_module = 1;
 }
