+++ tzrec/protos/module.proto
@@ -62,3 +62,15 @@ message MaskBlock{
     // the dim of hidden ffn layer
     required uint32 hidden_dim = 3;
 }
+
+message MaskNetModule {
+    // number of mask blocks
+    required uint32 n_mask_blocks = 1;
+    // mask block
+    required MaskBlock mask_block = 2;
+    // mlp layer on top of mask blocks
+    required MLP top_mlp = 3;
+    // use parallel or serial mask blocks
+    optional bool use_parallel = 4 [default=true];
+
+}
